training/sagemaker/requirements.txt
datasets
pytorch-lightningtorchmetrics 
transformers
